Organized project into class libraries (Controllers, Entities). Added dbcontext (Data) and a Migrations folder.
