+++ Entities/Settlement.cs
