+++ Raktar.sln
@@ -1,9 +1,13 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.13.35806.99 d17.13
+VisualStudioVersion = 17.13.35806.99
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Raktar", "Raktar\Raktar.csproj", "{A3B223B6-5B90-4E8D-B132-0C2DA48292B3}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Entities", "Entities\Entities.csproj", "{3354210D-EC10-4F61-8289-C9C55F6CD983}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Controllers", "Controllers\Controllers.csproj", "{A2A7868A-9F4E-4503-9A4A-C5A24698C200}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +19,14 @@ Global
 		{A3B223B6-5B90-4E8D-B132-0C2DA48292B3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A3B223B6-5B90-4E8D-B132-0C2DA48292B3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A3B223B6-5B90-4E8D-B132-0C2DA48292B3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3354210D-EC10-4F61-8289-C9C55F6CD983}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3354210D-EC10-4F61-8289-C9C55F6CD983}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3354210D-EC10-4F61-8289-C9C55F6CD983}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3354210D-EC10-4F61-8289-C9C55F6CD983}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A2A7868A-9F4E-4503-9A4A-C5A24698C200}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A2A7868A-9F4E-4503-9A4A-C5A24698C200}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A2A7868A-9F4E-4503-9A4A-C5A24698C200}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A2A7868A-9F4E-4503-9A4A-C5A24698C200}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
